tests/requirements.txt
# Test dependencies
pytest>=8.0.0
httpx>=0.27.0
websocket-client>=1.8.0
